requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
msal==1.24.1
google-auth==2.23.4
requests==2.31.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
email-validator==2.1.0
clerk-backend-sdk==0.9.0